feat(#2, #3): Implement restrictions in the API
Removed the set-cookie approach, relying on the external auth module to
do all the workload. At the moment there isn't any way to set a cookie
from there, but it could be added as a new element in the map inside
nrte_auth_value if required.

Also homogenized the interfaces so all the topics are always specified
via query string.

Closes #2, #3.